+++ Common/src/Socket/SocketWrapper.cpp
@@ -37,11 +37,11 @@ SocketWrapper SocketWrapper::accept() const {
 //-----------------------------------------------------------------------------
 // Sobrecarga de operadores
 
-ssize_t SocketWrapper::operator<<(uint16_t n) const {
+ssize_t SocketWrapper::operator<<(uint32_t n) const {
     return send((char*)&n, sizeof(n));
 }
 
-ssize_t SocketWrapper::operator>>(uint16_t& n) const {
+ssize_t SocketWrapper::operator>>(uint32_t& n) const {
     uint16_t received;
     ssize_t n_received;
     n_received = recv((char*)&received, sizeof(received));
